Fix RPM dependency conflict
MusiKernel statically links it's own Portaudio library to improve performance,
and drops a copy of the libportaudio.so binary into Python so that it can
use the same version for querying hardware information.  RPM was considering
the musikernel package as providing libportaudio.so , and not installing
the portaudio package for other packages that require it.