Install FFADO into the live DVD/USB image.
I have really mixed feelings about doing this, but people with Firewire
interfaces can at least be given a chance at using them...

I just dusted off my SaffirePro14 and gave it a whirl, attempted the
default Qjackctl settings that have no chance of working on the first try,
adjusted numerous settings, got it running with way too much latency,
and then was not able to get actual sound out of it despite mucking around
Qjackctl and FFADO's mixer for about 20 minutes.  BUT, the peak meters and
playback cursor in DAW-Next did indicate that it was outputting sound
somewhere, maybe just being muted by a volume slider buried somewhere in
the 2 prerequisite applications to run a fucking firewire interface in Linux.

I have a lingering suspicion that the forthcoming Windows and Mac ports of
MusiKernel will be much more popular with Firewire users.