Add a dh_make debian/ directory to ease making source debs for Launchpad
Having said that, Launchpad is absolutely horrid...  If it was properly
documented, people could be up-and-running in under a hour.  Instead, I
have to spend an entire night fumbling through the
"your package has been rejected" errors like I'm sure every developer
before me who ever used it had to do for their first PPA.

All of the GPG keys and signing and everything are all the right thing
to do, but the workflow and piss-poor documentation are as bad as it gets.

At the moment, I'm waiting to see if the web UI will delete all remnants
of my package so I can continue without bumping my version number just
for the sake of continuing to fumble through all of these silly error
messages.  I'm lead to believe that it will delete it within about 30
minutes, which means from here on out I get to fix up to 2 packaging bugs
per-hour until everything is fixed, because I refuse to bump version numbers
10 times in one day just to get through this.