Fix a horrible "optimization" KDE5 makes to Qt applications
It turns out that KDE5 will automatically make shortcuts for
random QAction's, meaning that if I autogenerate a menu full
of actions, and call one "my action 1", KDE, might decide to
rename it to "m&y action 1" to make the 'y' key a shortcut.
Adding random ampersands obviously breaks everything if the
actions were auto-generated and QAction.text() was being
relied on to store the name of the action.

So, don't rely on QAction.text(), store the name by dynamically
creating a variable in the QAction.  This fixes many, many
bugs under KDE5 (although generally, it was possible to not
actually encounter the bugs in real life if you didn't rely
on particular menu actions for your workflow)