Work around some odd behavior in Windows when the application closes
At some point during the main window closeEvent, it seems that
exit() is being called and the application never gets to the
if RESPAWN code path.

However, calling respawn early doesn't work in Linux because Linux
uses sockets for IPC by default and has a race condition between
close and listen, so continue the old way with Linux.