Work around the QMenu regression in Qt5.5.0
This removes the categorized nested plugin menu, which is very
unfortunate.  Qt5.5.0 is already in Fedora22 and Arch Linux, possibly
others.  Debian and Ubuntu are still using older versions.  The
regression has been reported to the Qt developers and assigned a
priority of "Important", but it's not clear when it will be fixed,
and when the fix will be released and placed into distro repositories,
hence the need to do something about it right now.